backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
